Release 0.23.0

+++ _scripts/install.sh
@@ -1,16 +1,23 @@
 #!/usr/bin/bash
+# set -x
 version=""
 version="--prerelease"
 cd src
-exes=`find . -name 'tr*.exe' | grep -v publish`
-for i in $exes
+dirs=`find .  -name net8.0 | fgrep 'bin/Release' | fgrep -v Generated | grep '^./tr' | fgrep -v publish | sort -u`
+for i in $dirs
 do
 	d=`echo $i | awk -F '/' '{print $2}'`
-	cd $d
+	pushd $i
+	if [ ! -f $d.dll ]
+	then
+		popd
+		continue
+	fi
+	popd
+	pushd $d
 	tool=$d
 	dotnet tool uninstall -g $tool $version > /dev/null 2>&1
-	echo dotnet tool install -g $tool $version 
 	dotnet tool install -g $tool $version > /dev/null 2>&1
 	$tool --version
-	cd ..
+	popd
 done
